+++ app/scripts/util/feature-detector.js
@@ -6,7 +6,8 @@ const FeatureDetector = {
     isWindows: navigator.platform.indexOf('Win') >= 0,
     isiOS: /iPad|iPhone|iPod/i.test(navigator.userAgent),
     isMobile: MobileRegex.test(navigator.userAgent) || screen.width < MinDesktopScreenWidth,
-
+    isPopup: !!((window.parent !== window.top) || window.opener),
+    isStandalone: !!(navigator.standalone || window.matchMedia('(display-mode: standalone)').matches),
     isBeta: window.location.href.toLowerCase().indexOf('beta.') > 0,
 
     actionShortcutSymbol: function(formatting) {