+++ app/scripts/auto-type/auto-type-filter.js
@@ -15,18 +15,15 @@ AutoTypeFilter.prototype.getEntries = function() {
         text: this.text,
         autoType: true
     };
-    let entries = this.appModel.getEntriesByFilter(filter);
-    if (!this.ignoreWindowInfo && this.hasWindowInfo()) {
-        this.prepareFilter();
-        entries = new EntryCollection(entries
-            .map(e => [e, this.getEntryRank(e)])
-            .filter(e => e[1])
-            .sort((x, y) => x[1] === y[1] ? x[0].title.localeCompare(y[0].title) : y[1] - x[1])
-            .map(p => p[0]));
-    } else {
-        entries.sortEntries('title');
+    this.prepareFilter();
+    let entries = this.appModel.getEntriesByFilter(filter)
+        .map(e => [e, this.getEntryRank(e)]);
+    if (!this.ignoreWindowInfo) {
+        entries = entries.filter(e => e[1]);
     }
-    return entries;
+    entries = entries.sort((x, y) => x[1] === y[1] ? x[0].title.localeCompare(y[0].title) : y[1] - x[1]);
+    entries = entries.map(p => p[0]);
+    return new EntryCollection(entries, {comparator: 'none'});
 };
 
 AutoTypeFilter.prototype.hasWindowInfo = function() {