+++ release-notes.md
@@ -3,6 +3,7 @@ Release notes
 ##### v1.5.0 (TBD)
 `+` plugins  
 `*` translations are available only as plugins  
+`+` mobile field editing improvements  
 `+` file path hint in recent files list  
 `+` cacheConfigSettings config option  
 `+` keyboard-accessible autocomplete