+++ release-notes.md
@@ -6,6 +6,7 @@ Release notes
 `*` Dropbox API V2  
 `+` entry templates  
 `+` support cloud providers in iOS homescreen apps  
+`+` auto-type improvements  
 `+` mobile field editing improvements  
 `+` file path hint in recent files list  
 `+` cacheConfigSettings config option