+++ release-notes.md
@@ -16,6 +16,7 @@ Release notes
 `+` showOnlyFilesFromConfig config option  
 `+` mac os window style options  
 `+` lock on hide mac os hide  
+`+` release distribution improvements  
 `-` removed support of devices without clipboard api  
 
 ##### v1.4.1 (2017-03-26)